add ExtractConfig, TimestampColumn interface in typescript, and refactor GroupReducer to its own file
