+++ app/src/FileStatReducer.tsx
